requirements/common.txt
Django==4.0.1
attrs==21.4.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3